cleaned up files, added Behoulder, started items
- made a bunch of folders for each asset
- made the default enemy into the "Behoulder"
- started item manager to put effect and value into to create easy stat up items; will switch to array for multiple stats

- next: invul frames, stat visuals, audio, pausing/settings/save settings, item/enemy spawning

+++ Assets/Scripts/PlayerMovement.cs
@@ -9,7 +9,7 @@ public class PlayerMovement : MonoBehaviour
     private float lastHorizontal;
     private float lastVertical;
     private float speed;
-    private int health;
+    private float health;
     [SerializeField] private Rigidbody2D rb;
     [SerializeField] private GameObject pickaxePrefab;
     [SerializeField] private Transform firePoint;
@@ -133,7 +133,7 @@ public class PlayerMovement : MonoBehaviour
     {
         if (collision.gameObject.CompareTag("Enemy"))
         {
-            health -= 1;
+            health -= 1f;
         }
     }
 }
